backend/requirements.txt
# ============================================
# GLOBAL MOBILITY INTELLIGENCE PLATFORM
# Backend Dependencies - Compatible Versions
# ============================================

# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0

# HTTP Client (for direct API calls - avoids SDK version conflicts)
httpx>=0.28.0

# CrewAI & LLM (Optional - will fall back to direct API if not working)
crewai>=0.86.0
crewai-tools>=0.17.0
openai>=1.50.0
groq>=0.4.1

# LangChain
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0

# RAG & Vector Store
chromadb>=0.4.0
tiktoken>=0.8.0
sentence-transformers>=2.2.0  # Free local embeddings

# Utilities
tenacity>=9.0.0
aiohttp>=3.11.0

# Additional Dependencies
numpy>=1.26.0
regex>=2024.0.0
python-multipart>=0.0.12
jinja2>=3.1.0
